// src/types/main/main.ts
export interface FreightMonthlySubsidyPaymentStatus {
    type: 'freight';
    year?: string   // 청구월
    opsAmt: string // 지급금액
    crtrYm: string //청구년월
}

// 버스 월별 보조금 지급현황을 조회             [] 페이징
//  /fsm/mai/main/bs/getOpsMth      파라미터 ex year : 2024
export interface BusMonthlySubsidyPaymentStatus {
    type: 'bus';
    year?: string   // 청구월
    opsAmt: string // 지급금액
    crtrYm: string //청구년월
}

// 택시 월별 보조금 지급현황을 조회                 [] 페이징
// /fsm/mai/main/tx/getOpsMthTx     파라미터 ex year 2024
export interface TaxiMonthlySubsidyPaymentStatus {
    type: 'taxi';
    year?: string   // 청구월
    opsAmt: string // 지급금액
    crtrYm: string //청구년월
}